style: moved test files to folders of tested modules
